connection panel wrapper component for connection components, added proptypes
