--- app/ui/components/ConnectionInfo.css
@@ -1,23 +1,0 @@
-.connect{
-  padding:0.5rem 0.5rem;
-}
-
-.msg{
-  font-family:'sans';
-  font-size: 0.75rem;
-  line-height: 1rem;
-  padding-bottom: 1rem
-}
-
-.status{
-	text-transform: uppercase;
-}
-
-.location{
-	font-family: 'sans';
-	line-height: 1rem;
-}
-
-.connect_btn{
-  width:100%;
-}
